% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ls.R
\name{read_h5ls}
\alias{read_h5ls}
\title{read_h5ls}
\usage{
read_h5ls(filename)
}
\arguments{
\item{filename}{Name of the h5ls output file to be read.}
}
\value{
A tibble
}
\description{
Read the file 'filename', expected to be the output of "h5ls -vr <file>".
Return a tibble with one row per dataset found in the file.
}
